Rename gumbo.h to nokogiri_gumbo.h
In some cases, building Nokogiri can pick up a system-installed
libgumbo. By renaming the one external header, this should no longer
happen.

Fixes: #2464
