+++ gumbo-parser/src/tokenizer.h
@@ -7,7 +7,7 @@
 #include <stdbool.h>
 #include <stddef.h>
 
-#include "gumbo.h"
+#include "nokogiri_gumbo.h"
 #include "token_type.h"
 #include "tokenizer_states.h"
