+++ gumbo-parser/src/util.c
@@ -19,7 +19,7 @@
 #include <stdlib.h>
 #include <string.h>
 #include "util.h"
-#include "gumbo.h"
+#include "nokogiri_gumbo.h"
 
 void* gumbo_alloc(size_t size) {
   void* ptr = malloc(size);
